requirements.txt
selenium==4.40.0
requests==2.32.5
pytest==9.0.2
allure-pytest==2.15.3
webdriver-manager==4.0.2
python-dotenv==1.2.1
pydantic==2.5.0
typing-extensions==4.15.0